Account for coinbase + holding BTCZ address (#2)
Since pools donate directly to the donation address, it receives
unshielded funds. This means we have to transfer the funds to a
zaddress, then back to a taddress.

To keep this process from getting too complicated, we're using the
taddress previously published everywhere as the "receiving" address,
with funds being migrated to a holding address which people shouldn't be
donating to, so that we don't have to shield back and forth.

This change updates the site to account for both addresses.